+++ bc/recruitment_api/management/commands/import_jobs.py
@@ -1,6 +1,8 @@
+from django.core.files.base import ContentFile
 from django.core.management.base import BaseCommand
 from django.utils.timezone import now
 
+from bc.documents.models import CustomDocument
 from bc.recruitment.models import TalentLinkJob
 from bc.recruitment_api.client import get_client
 from bc.recruitment_api.utils import update_job_from_ad
@@ -23,6 +25,7 @@ class Command(BaseCommand):
         results = True
         num_updated = 0
         num_created = 0
+        doc_imported = 0
         errors = []
         while results:
             self.stdout.write(f"Fetching page {page}")
@@ -61,9 +64,43 @@ class Command(BaseCommand):
                         else:
                             num_updated += 1
 
+                        # Fetch attachments via a different call
+                        try:
+                            # This will return list of attachments with
+                            #   'content', 'description', 'fileName', 'id', 'mimeType'
+                            attachments_response = client.service.getAttachments(
+                                job.talentlink_id
+                            )
+                            for attachment in attachments_response:
+                                if attachment["id"] and attachment["fileName"]:
+                                    doc, created = CustomDocument.objects.get_or_create(
+                                        talentlink_attachment_id=attachment["id"]
+                                    )
+                                    if created:
+                                        doc.title = (
+                                            attachment["description"]
+                                            or attachment["fileName"].split(".")[0]
+                                        )
+                                        doc.file = ContentFile(
+                                            attachment["content"],
+                                            name=attachment["fileName"],
+                                        )
+                                        doc.save()
+                                        doc_imported += 1
+
+                                    job.attachments.add(doc)
+                                    job.save()
+                        except Exception as e:
+                            msg = (
+                                f"Error occurred while importing attachments for job {ad['id']}:\n"
+                                + str(e)
+                            )
+                            errors.append(msg)
+
         self.stdout.write("No more results")
         self.stdout.write(f"{num_updated} existing jobs updated")
         self.stdout.write(f"{num_created} new jobs created")
+        self.stdout.write(f"{doc_imported} new documents imported")
         self.stdout.write(self.style.ERROR(f"{len(errors)} errors"))
         for error in errors:
             self.stdout.write(self.style.ERROR(msg))