+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/8.0...9.0
 - 129 Jobs postcode search
 - 169 Adjust position of form field help text
 - Forms integration with Aptean Respond API
+- Fix an issue with CI deployments
 
 ## 8.0 (2020-03-30)